requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==1.11.12
django-rest-swagger==2.1.2
djangorestframework==3.8.2
drf-openapi==1.3.0
drf-yasg==1.6.1
flex==6.13.1
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe==1.0
mock==2.0.0
openapi-codec==1.3.2
pbr==4.0.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rfc3987==1.3.7
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.37
selenium==3.11.0
simplejson==3.13.2
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
